requirements.txt
beautifulsoup4==4.12.3
pinecone==5.0.0
pinecone-client==5.0.0
langchain==0.2.15
langchain-community==0.2.15
sentence-transformers==3.0.0
python-dotenv==1.0.0
aiohttp==3.10.0
aiosqlite==0.20.0
streamlit==1.37.1
langchain-google-genai==1.0.9
langchain-huggingface==0.0.3
langchain-pinecone==0.1.3
google-generativeai==0.7.2